requirements.txt
SpeechRecognition==3.14.3
pyttsx3==2.99
psutil==7.0.0
Pillow==11.3.0
pyautogui==0.9.54
PyAudio==0.2.14 
autopep8==2.3.1
flake8==7.1.1
pylint==3.3.1